@@ -1,56 +1,52 @@
 {
-    "name": "my-lib",
-    "version": "0.0.0",
-    "scripts": {
-        "dev": "vite",
-        "build": "tsc && vite build",
-        "preview": "vite preview",
-        "prepack": "json -f package.json -I -e \"delete this.devDependencies; delete this.dependencies\"",
-        "storybook": "start-storybook -p 6006",
-        "build-storybook": "build-storybook"
-    },
-    "dependencies": {
-        "react": "17.0.2",
-        "react-dom": "17.0.2",
-        "styled-components": "5.3.5"
-    },
-    "devDependencies": {
-        "@babel/core": "7.17.8",
-        "@storybook/addon-a11y": "6.5.0-alpha.50",
-        "@storybook/addon-actions": "6.5.0-alpha.50",
-        "@storybook/addon-docs": "6.5.0-alpha.50",
-        "@storybook/addon-essentials": "6.5.0-alpha.50",
-        "@storybook/addon-links": "6.5.0-alpha.50",
-        "@storybook/react": "6.5.0-alpha.49",
-        "@types/node": "17.0.23",
-        "@types/react": "17.0.43",
-        "@types/react-dom": "17.0.14",
-        "@types/styled-components": "5.1.24",
-        "@vitejs/plugin-react": "1.2.0",
-        "acorn-jsx": "5.3.2",
-        "babel-loader": "8.2.4",
-        "json": "11.0.0",
-        "storybook-builder-vite": "0.1.21",
-        "typescript": "4.6.3",
-        "vite": "2.8.6",
-        "vite-plugin-dts": "1.0.3"
-    },
-    "license": "UNLICENSED",
-    "peerDependencies": {
-        "react": "16.8.0 || 17.x",
-        "react-dom": "16.8.0 || 17.x",
-        "styled-components": "5.0.0"
-    },
-    "files": [
-        "dist"
-    ],
-    "main": "./dist/my-lib.umd.js",
-    "module": "./dist/my-lib.es.js",
-    "types": "./dist/index.d.ts",
-    "exports": {
-        ".": {
-            "import": "./dist/my-lib.es.js",
-            "require": "./dist/my-lib.umd.js"
-        }
+  "name": "my-lib",
+  "version": "0.0.0",
+  "scripts": {
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "preview": "vite preview",
+    "prepack": "json -f package.json -I -e \"delete this.devDependencies; delete this.dependencies\"",
+    "storybook": "start-storybook -p 6006",
+    "build-storybook": "build-storybook"
+  },
+  "dependencies": {
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "styled-components": "^5.3.5"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.18.9",
+    "@storybook/addon-a11y": "^6.5.9",
+    "@storybook/addon-actions": "^6.5.9",
+    "@storybook/addon-docs": "^6.5.9",
+    "@storybook/addon-essentials": "^6.5.9",
+    "@storybook/addon-links": "^6.5.9",
+    "@storybook/builder-vite": "^0.2.0",
+    "@storybook/react": "^6.5.9",
+    "@types/node": "^18.0.6",
+    "@types/react": "^18.0.15",
+    "@types/react-dom": "^18.0.6",
+    "@types/styled-components": "^5.1.25",
+    "@vitejs/plugin-react": "^2.0.0",
+    "acorn-jsx": "^5.3.2",
+    "babel-loader": "^8.2.5",
+    "json": "^11.0.0",
+    "typescript": "^4.7.4",
+    "vite": "^3.0.2",
+    "vite-plugin-dts": "^1.4.0"
+  },
+  "license": "UNLICENSED",
+  "peerDependencies": {},
+  "files": [
+    "dist"
+  ],
+  "main": "./dist/my-lib.umd.js",
+  "module": "./dist/my-lib.es.js",
+  "types": "./dist/index.d.ts",
+  "exports": {
+    ".": {
+      "import": "./dist/my-lib.es.js",
+      "require": "./dist/my-lib.umd.js"
     }
+  }
 }
