@@ -7,16 +7,15 @@
         "preview": "vite preview"
     },
     "dependencies": {
-        "my-lib": "*",
-        "react": "17.0.2",
-        "react-dom": "17.0.2",
-        "styled-components": "5.3.5"
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
+        "styled-components": "^5.3.5"
     },
     "devDependencies": {
-        "@types/react": "17.0.43",
-        "@types/react-dom": "17.0.14",
-        "@vitejs/plugin-react": "1.2.0",
-        "typescript": "4.6.3",
-        "vite": "2.8.6"
+        "@types/react": "^18.0.15",
+        "@types/react-dom": "^18.0.6",
+        "@vitejs/plugin-react": "^2.0.0",
+        "typescript": "^4.7.4",
+        "vite": "^3.0.2"
     }
 }
