@@ -1,7 +1,9 @@
 import react from '@vitejs/plugin-react';
-import path from 'node:path';
-import { defineConfig } from 'vite';
+import * as path from 'node:path';
+import {defineConfig} from 'vite';
 import dts from 'vite-plugin-dts';
+
+const config = require('./package.json');
 
 export default defineConfig({
     plugins: [
@@ -14,9 +16,9 @@ export default defineConfig({
         sourcemap: true,
         lib: {
             entry: path.resolve(__dirname, 'src/lib/index.ts'),
-            name: 'MyLib',
+            name: config.name,
             formats: ['es', 'umd'],
-            fileName: (format) => `my-lib.${format}.js`,
+            fileName: (format) => `${config.name}.${format}.js`,
         },
         rollupOptions: {
             external: ['react', 'react-dom', 'styled-components'],
